Fix anti falsing detection
The falsing detection stopped listening for touches
when the lockscreen showed without turning the screen off.

This happens e.g. when SystemUI crashes, when switching users,
when ending lock-to-task.

As a result, the falsing detection usually got stuck on a bad
value and without further touches continued preventing all input.

Bug: 27227578
Change-Id: I6717c04b0271e4d2578b88fc584b6ad46492e497