HdmiControlService does not broadcast CEC <standby> message on shutdown
When the STB is shutting down the CEC <standby> message should be
broadcast to all CEC devices to place them in to standby. This is
not happening because the ACTION_SHUTDOWN Intent has not been
added to the list of Intent filters that the HdmiControlService
should be listening for.

This one-line change adds the ACTION_SHUTDOWN Intent filter so that
when the STB is shutting down (e.g. via a long press of the POWER
button), the CEC <standby> message can be broadcast to all devices.
This allows a connected HDMI TV to be automatically placed in to
standby.

Bug: 27596547
Change-Id: I4cdd2eaaad86e98bdf86c74fb0c377eb716f3c83