Do not set docked divider as IME target.
If we do this, we will fail to adjust the proper IME target
to make it visible. Accomplish relative ordering of IME and
Docked Divider in WindowLayersController instead.

We need to take care that adjustSpecialWindows won't push windows
down if they are already positioned above the highest application
layer. We also take care to not adjust the IME if the docked divider
isn't really visible.

Bug: 26387930
Change-Id: I26ca36c4f7ecf9d97f44e15c67df82b8154a169c